Fix flight duration display
